+++ test/test-applications/integrations/TestApplication.SqlClient.System.NetFramework/Program.cs
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 using System.Data.SqlClient;
+using System.Diagnostics;
 using TestApplication.Shared;
 
 namespace TestApplication.SqlClient.System;
@@ -30,6 +31,19 @@ public class Program
         using (var connection = new SqlConnection(connectionString))
         {
             await ExecuteAsyncCommands(connection);
+        }
+
+        // The "LONG_RUNNING" environment variable is used by tests that access/receive
+        // data that takes time to be produced.
+        var longRunning = Environment.GetEnvironmentVariable("LONG_RUNNING");
+        if (longRunning == "true")
+        {
+            // In this case it is necessary to ensure that the test has a chance to read the
+            // expected data, only by keeping the application alive for some time that can
+            // be ensured. Anyway, tests that set "LONG_RUNNING" env var to true are expected
+            // to kill the process directly.
+            Console.WriteLine("LONG_RUNNING is true, waiting for process to be killed...");
+            Process.GetCurrentProcess().WaitForExit();
         }
     }